[sw/sca] Add side-channel analysis support library (sca)
This change introduces a library that provides various functionality
that is expected to be shared across different SCA targets such as:
- Peripheral initialization (pinmux, UART, GPIO, and timer),
- Trigger handling,
- IRQ handling, and
- Sleeping during encryption operations.
