+++ sw/device/sca/aes_serial/aes_serial.c
@@ -3,27 +3,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "sw/device/lib/aes.h"
-#include "sw/device/lib/arch/device.h"
-#include "sw/device/lib/dif/dif_gpio.h"
-#include "sw/device/lib/dif/dif_rv_timer.h"
-#include "sw/device/lib/dif/dif_uart.h"
-#include "sw/device/lib/handler.h"
-#include "sw/device/lib/irq.h"
-#include "sw/device/lib/pinmux.h"
 #include "sw/device/lib/runtime/hart.h"
 #include "sw/device/lib/runtime/log.h"
 #include "sw/device/lib/runtime/print.h"
-
-#include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
+#include "sw/device/sca/aes_serial/sca.h"
 
 // This program implements the ChipWhisperer simple serial protocol version 1.1.
 // Only set key ('k'), encrypt ('p') and version ('v') commands are implemented.
 // Encryption is done in AES-ECB-128 mode.
 // See https://wiki.newae.com/SimpleSerial for details.
-
-static dif_gpio_t gpio;
-static dif_rv_timer_t timer;
-static dif_uart_t uart;
 
 enum {
   kAesKeySize = 16,
@@ -38,16 +26,6 @@ enum {
    */
   kIbexAesSleepCycles = 200,
   /**
-   * GPIO capture trigger values.
-   */
-  kGpioCaptureTriggerHigh = 0x08200,
-  kGpioCaptureTriggerLow = 0x00000,
-  /**
-   * RV timer settings.
-   */
-  kRvTimerComparator = 0,
-  kRvTimerHart = kTopEarlgreyPlicTargetIbex0,
-  /**
    * UART constants.
    */
   kUartMaxRxSizeText = 128,
@@ -58,28 +36,13 @@ enum {
   kSimpleSerialProtocolVersion = 1,
 };
 
+static const dif_uart_t *uart;
+
 // Simple serial status codes
 typedef enum simple_serial_result {
   kSimpleSerialOk = 0,
   kSimpleSerialError = 1
 } simple_serial_result_t;
-
-// TODO: Remove once there is a CHECK version that does not require test
-// library dependencies.
-#define CHECK(condition, ...)                             \
-  do {                                                    \
-    if (!(condition)) {                                   \
-      /* NOTE: because the condition in this if           \
-         statement can be statically determined,          \
-         only one of the below string constants           \
-         will be included in the final binary.*/          \
-      if (GET_NUM_VARIABLE_ARGS(_, ##__VA_ARGS__) == 0) { \
-        LOG_ERROR("CHECK-fail: " #condition);             \
-      } else {                                            \
-        LOG_ERROR("CHECK-fail: " __VA_ARGS__);            \
-      }                                                   \
-    }                                                     \
-  } while (false)
 
 /** Returns kSimpleSerialError if the condition evaluates to false */
 #define SS_CHECK(condition)      \
@@ -88,116 +51,6 @@ typedef enum simple_serial_result {
       return kSimpleSerialError; \
     }                            \
   } while (false)
-
-/**
- * Initializes the UART peripheral.
- */
-static void sca_init_uart() {
-  (void)dif_uart_init(
-      (dif_uart_params_t){
-          .base_addr = mmio_region_from_addr(TOP_EARLGREY_UART_BASE_ADDR),
-      },
-      &uart);
-  (void)dif_uart_configure(&uart, (dif_uart_config_t){
-                                      .baudrate = kUartBaudrate,
-                                      .clk_freq_hz = kClockFreqPeripheralHz,
-                                      .parity_enable = kDifUartToggleDisabled,
-                                      .parity = kDifUartParityEven,
-                                  });
-  dif_uart_fifo_reset(&uart, kDifUartFifoResetAll);
-  base_uart_stdout(&uart);
-}
-
-/**
- * Initializes the GPIO peripheral.
- */
-static void sca_init_gpio() {
-  dif_gpio_params_t gpio_params = {
-      .base_addr = mmio_region_from_addr(TOP_EARLGREY_GPIO_BASE_ADDR)};
-  (void)dif_gpio_init(gpio_params, &gpio);
-  (void)dif_gpio_output_set_enabled_all(&gpio, kGpioCaptureTriggerHigh);
-}
-
-/**
- * Initializes the timer peripheral.
- */
-static void sca_init_timer() {
-  (void)dif_rv_timer_init(
-      mmio_region_from_addr(TOP_EARLGREY_RV_TIMER_BASE_ADDR),
-      (dif_rv_timer_config_t){.hart_count = 1, .comparator_count = 1}, &timer);
-  dif_rv_timer_tick_params_t tick_params;
-  (void)dif_rv_timer_approximate_tick_params(kClockFreqPeripheralHz,
-                                             kClockFreqCpuHz, &tick_params);
-  (void)dif_rv_timer_set_tick_params(&timer, kRvTimerHart, tick_params);
-  (void)dif_rv_timer_irq_enable(&timer, kRvTimerHart, kRvTimerComparator,
-                                kDifRvTimerEnabled);
-  irq_timer_ctrl(true);
-  irq_global_ctrl(true);
-}
-
-/**
- * Timer IRQ handler.
- *
- * Disables the counter and clears pending interrupts.
- */
-void handler_irq_timer(void) {
-  // Return values of below functions are ignored to improve capture
-  // performance.
-  (void)dif_rv_timer_counter_set_enabled(&timer, kRvTimerHart,
-                                         kDifRvTimerDisabled);
-  (void)dif_rv_timer_irq_clear(&timer, kRvTimerHart, kRvTimerComparator);
-}
-
-/**
- * Initializes the peripherals (pinmux, uart, gpio, and timer) used by SCA code.
- */
-static void sca_init(void) {
-  pinmux_init();
-  sca_init_uart();
-  sca_init_gpio();
-  sca_init_timer();
-}
-
-/**
- * Sets capture trigger high.
- *
- * The actual trigger signal used for capture is a combination (logical AND) of:
- * - GPIO15 enabled here, and
- * - the busy signal of the AES unit. This signal will go high 40 cycles
- *   after aes_manual_trigger() is executed below and remain high until
- *   the operation has finished.
- */
-static void sca_set_trigger_high() {
-  (void)dif_gpio_write_all(&gpio, kGpioCaptureTriggerHigh);
-}
-
-/**
- * Sets capture trigger low.
- */
-static void sca_set_trigger_low() {
-  (void)dif_gpio_write_all(&gpio, kGpioCaptureTriggerLow);
-}
-
-/**
- * Functions called by `sca_call_and_sleep()` must conform to this prototype.
- */
-typedef void (*sca_callee)(void);
-
-static void sca_call_and_sleep(sca_callee callee, uint32_t sleep_cycles) {
-  // Start timer to wake Ibex after AES is done.
-  uint64_t current_time;
-  // Return values of below functions are ignored to improve capture
-  // performance.
-  (void)dif_rv_timer_counter_read(&timer, kRvTimerHart, &current_time);
-  (void)dif_rv_timer_arm(&timer, kRvTimerHart, kRvTimerComparator,
-                         current_time + sleep_cycles);
-  (void)dif_rv_timer_counter_set_enabled(&timer, kRvTimerHart,
-                                         kDifRvTimerEnabled);
-
-  callee();
-
-  wait_for_interrupt();
-}
 
 /**
  * Convert `from` binary `to` hex.
@@ -375,6 +228,7 @@ static void simple_serial_handle_command(const aes_cfg_t *aes_cfg,
 
 int main(void) {
   sca_init();
+  sca_get_uart(&uart);
 
   aes_cfg_t aes_cfg;
   aes_cfg.key_len = kAes128;
@@ -386,7 +240,7 @@ int main(void) {
   size_t pos = 0;
   while (true) {
     size_t chars_read;
-    if (dif_uart_bytes_receive(&uart, 1, (uint8_t *)&text[pos], &chars_read) !=
+    if (dif_uart_bytes_receive(uart, 1, (uint8_t *)&text[pos], &chars_read) !=
             kDifUartOk ||
         chars_read == 0) {
       usleep(50);
